Rncalation (ES): Address PR review comments
