+++ src/es/nartag/src/eu/kanade/tachiyomi/extension/es/nartag/Filters.kt
@@ -1,0 +1,44 @@
+package eu.kanade.tachiyomi.extension.es.nartag
+
+import eu.kanade.tachiyomi.source.model.Filter
+
+class TypeFilter : Filter.Select<String>("Tipo", arrayOf("Todos", "Manga", "Manhwa", "Manhua", "Novel", "Other"))
+
+class StatusFilter : Filter.Select<String>("Estado", arrayOf("Todos", "Ongoing", "Completed", "Hiatus", "Cancelled"))
+
+class SortFilter : Filter.Select<SortOption>("Ordenar por", sortOptions)
+
+class SortOption(val name: String, val value: String) {
+    override fun toString(): String = name
+}
+
+class GenreFilter(genres: List<Genre>) : Filter.Group<Genre>("Géneros", genres)
+class Genre(name: String) : Filter.CheckBox(name)
+
+val sortOptions = arrayOf(
+    SortOption("Más reciente", "latest"),
+    SortOption("Actualizado", "updated"),
+    SortOption("Más visto", "views"),
+    SortOption("Mejor valorado", "rating"),
+    SortOption("A-Z", "title"),
+)
+
+val genresList = listOf(
+    "Acción",
+    "Adventure",
+    "Aventura",
+    "ciencia ficción",
+    "Comedia",
+    "Cultivación",
+    "Drama",
+    "Fantasia",
+    "Fantasía",
+    "Harem",
+    "Love",
+    "Manhua",
+    "Reencarnacion",
+    "Romance",
+    "Sistema",
+    "Supernatural",
+    "+15",
+)
